+++ graphqlcodegen-maven-plugin/src/main/java/io/github/deweyjose/graphqlcodegen/services/SchemaFileService.java
@@ -1,4 +1,4 @@
-package io.github.deweyjose.graphqlcodegen;
+package io.github.deweyjose.graphqlcodegen.services;
 
 import java.io.File;
 import java.io.IOException;
