+++ graphqlcodegen-maven-plugin/src/test/java/io/github/deweyjose/graphqlcodegen/services/SchemaManifestServiceTest.java
@@ -1,4 +1,4 @@
-package io.github.deweyjose.graphqlcodegen;
+package io.github.deweyjose.graphqlcodegen.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
